# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DDpay https://github.com/zkeai/DDPay",
        "title": "DDPay API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:2900",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/subsite/config": {
            "get": {
                "description": "获取分站JSON格式配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subsite"
                ],
                "summary": "获取分站JSON配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分站ID",
                        "name": "subsite_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "保存分站JSON格式配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subsite"
                ],
                "summary": "保存分站JSON配置",
                "parameters": [
                    {
                        "description": "保存分站JSON配置请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/subsite/create": {
            "post": {
                "description": "创建一个新的分站",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subsite"
                ],
                "summary": "创建分站",
                "parameters": [
                    {
                        "description": "创建分站请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateSubsiteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/subsite/delete": {
            "delete": {
                "description": "删除指定的分站",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subsite"
                ],
                "summary": "删除分站",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分站ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/subsite/info": {
            "get": {
                "description": "获取分站详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subsite"
                ],
                "summary": "获取分站信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/subsite/list": {
            "get": {
                "description": "获取所有分站列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subsite"
                ],
                "summary": "获取分站列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/subsite/update": {
            "put": {
                "description": "更新分站信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subsite"
                ],
                "summary": "更新分站",
                "parameters": [
                    {
                        "description": "更新分站请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateSubsiteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "conf.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.CreateSubsiteReq": {
            "type": "object",
            "required": [
                "commission_rate",
                "name",
                "subdomain"
            ],
            "properties": {
                "commission_rate": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                },
                "name": {
                    "type": "string"
                },
                "subdomain": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateSubsiteReq": {
            "type": "object",
            "required": [
                "commission_rate",
                "id",
                "name",
                "subdomain"
            ],
            "properties": {
                "commission_rate": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                },
                "description": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "subdomain": {
                    "type": "string"
                },
                "theme": {
                    "type": "string"
                }
            }
        },
...
    }
}